requirements.txt
Flask==2.0.3
scikit-learn==1.0.2
joblib==1.1.0
mysql-connector-python==8.0.28
pandas==1.3.5
